requirements.txt
google-adk>=0.2.0
spacy>=3.4.0
sentence-transformers>=2.2.0
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0

